2016/9/15 新增 数据类型binary的size属性

+++ Modifier/ValueStringMap.cs
@@ -62,7 +62,7 @@ namespace Modifier
             }
             return key;
         }
-        public string GetValue(int key)
+        public string GetValue(long key)
         {
             string value = "";
             foreach (var unit in GetMap())
